fix: downgrade react and react-dom version from 17.0.0 to 16.14.0
Version 17.0.0 of `react` and `react-dom` was breaking `storybook`. The sidebar does not work in production mode and the error message "Invalid hook call" has also been thrown.

After running the command: `yarn why react` two versions of React were found.

To fix this it was necessary to downgrade from version 17.0.0 to version 16.14.0